requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
